Fixed a logic bug with the implementation of variable step frame rates where bullet speed was being calculated based on elapse game time at time of bullet creation which makes no sense.
speed of the bullet is now calculated at time of bullet creation and then it's position during each update cycle is the updated based on this speed and varied based on elapse game time since last update.

This issue was causing the occasional slow bullet to be spawn since during shooting, when the game is running at a faster frame rate the elapse game time number is smaller at bullet spawn time which results in a very slow bullet being spawned and who's update method was never implemented correctly to use a variable step frame rate anyway.